HrAgent/AgentBase/requirements.txt
fastapi[standard]
uvicorn[standard]
streamlit
google-genai
python-dotenv
requests
chromadb
sentence-transformers
transformers
torch